move NodeTypesWithDBAdapter to reth-node-types
