+++ crates/node/api/Cargo.toml
@@ -14,8 +14,6 @@ workspace = true
 # reth
 reth-evm.workspace = true
 reth-provider.workspace = true
-reth-chainspec.workspace = true
-reth-db-api.workspace = true
 reth-engine-primitives.workspace = true
 reth-transaction-pool.workspace = true
 reth-payload-builder.workspace = true
