+++ crates/node/api/src/node.rs
@@ -2,91 +2,16 @@
 
 use std::marker::PhantomData;
 
-use reth_chainspec::EthChainSpec;
-use reth_db_api::{
-    database::Database,
-    database_metrics::{DatabaseMetadata, DatabaseMetrics},
-};
 use reth_evm::execute::BlockExecutorProvider;
 use reth_network_api::FullNetwork;
-use reth_node_types::{NodePrimitives, NodeTypes, NodeTypesWithDB, NodeTypesWithEngine};
+use reth_node_types::{NodeTypesWithDB, NodeTypesWithEngine};
 use reth_payload_builder::PayloadBuilderHandle;
 use reth_provider::FullProvider;
 use reth_rpc_eth_api::EthApiTypes;
 use reth_tasks::TaskExecutor;
 use reth_transaction_pool::TransactionPool;
 
-use crate::{ConfigureEvm, EngineTypes};
-
-/// A [`NodeTypes`] type builder.
-#[derive(Default, Debug)]
-pub struct AnyNodeTypes<P = (), C = ()>(PhantomData<P>, PhantomData<C>);
-
-impl<P, C> AnyNodeTypes<P, C> {
-    /// Sets the `Primitives` associated type.
-    pub const fn primitives<T>(self) -> AnyNodeTypes<T, C> {
-        AnyNodeTypes::<T, C>(PhantomData::<T>, PhantomData::<C>)
-    }
-
-    /// Sets the `ChainSpec` associated type.
-    pub const fn chain_spec<T>(self) -> AnyNodeTypes<P, T> {
-        AnyNodeTypes::<P, T>(PhantomData::<P>, PhantomData::<T>)
-    }
-}
-
-impl<P, C> NodeTypes for AnyNodeTypes<P, C>
-where
-    P: NodePrimitives + Send + Sync + Unpin + 'static,
-    C: EthChainSpec,
-{
-    type Primitives = P;
-    type ChainSpec = C;
-}
-
-/// A [`NodeTypesWithEngine`] type builder.
-#[derive(Default, Debug)]
-pub struct AnyNodeTypesWithEngine<P = (), E = (), C = ()> {
-    /// Embedding the basic node types.
-    base: AnyNodeTypes<P, C>,
-    /// Phantom data for the engine.
-    _engine: PhantomData<E>,
-}
-
-impl<P, E, C> AnyNodeTypesWithEngine<P, E, C> {
-    /// Sets the `Primitives` associated type.
-    pub const fn primitives<T>(self) -> AnyNodeTypesWithEngine<T, E, C> {
-        AnyNodeTypesWithEngine { base: self.base.primitives::<T>(), _engine: PhantomData }
-    }
-
-    /// Sets the `Engine` associated type.
-    pub const fn engine<T>(self) -> AnyNodeTypesWithEngine<P, T, C> {
-        AnyNodeTypesWithEngine { base: self.base, _engine: PhantomData::<T> }
-    }
-
-    /// Sets the `ChainSpec` associated type.
-    pub const fn chain_spec<T>(self) -> AnyNodeTypesWithEngine<P, E, T> {
-        AnyNodeTypesWithEngine { base: self.base.chain_spec::<T>(), _engine: PhantomData }
-    }
-}
-
-impl<P, E, C> NodeTypes for AnyNodeTypesWithEngine<P, E, C>
-where
-    P: NodePrimitives + Send + Sync + Unpin + 'static,
-    E: EngineTypes + Send + Sync + Unpin,
-    C: EthChainSpec,
-{
-    type Primitives = P;
-    type ChainSpec = C;
-}
-
-impl<P, E, C> NodeTypesWithEngine for AnyNodeTypesWithEngine<P, E, C>
-where
-    P: NodePrimitives + Send + Sync + Unpin + 'static,
-    E: EngineTypes + Send + Sync + Unpin,
-    C: EthChainSpec,
-{
-    type Engine = E;
-}
+use crate::ConfigureEvm;
 
 /// A helper trait that is downstream of the [`NodeTypesWithEngine`] trait and adds stateful
 /// components to the node.
@@ -97,57 +22,6 @@ pub trait FullNodeTypes: Send + Sync + Unpin + 'static {
     type Types: NodeTypesWithDB;
     /// The provider type used to interact with the node.
     type Provider: FullProvider<Self::Types>;
-}
-
-/// An adapter type combining [`NodeTypes`] and db into [`NodeTypesWithDB`].
-#[derive(Debug)]
-pub struct NodeTypesWithDBAdapter<Types, DB> {
-    types: PhantomData<Types>,
-    db: PhantomData<DB>,
-}
-
-impl<Types, DB> NodeTypesWithDBAdapter<Types, DB> {
-    /// Create a new adapter with the configured types.
-    pub fn new() -> Self {
-        Self { types: Default::default(), db: Default::default() }
-    }
-}
-
-impl<Types, DB> Default for NodeTypesWithDBAdapter<Types, DB> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<Types, DB> Clone for NodeTypesWithDBAdapter<Types, DB> {
-    fn clone(&self) -> Self {
-        Self { types: self.types, db: self.db }
-    }
-}
-
-impl<Types, DB> NodeTypes for NodeTypesWithDBAdapter<Types, DB>
-where
-    Types: NodeTypes,
-    DB: Send + Sync + Unpin + 'static,
-{
-    type Primitives = Types::Primitives;
-    type ChainSpec = Types::ChainSpec;
-}
-
-impl<Types, DB> NodeTypesWithEngine for NodeTypesWithDBAdapter<Types, DB>
-where
-    Types: NodeTypesWithEngine,
-    DB: Send + Sync + Unpin + 'static,
-{
-    type Engine = Types::Engine;
-}
-
-impl<Types, DB> NodeTypesWithDB for NodeTypesWithDBAdapter<Types, DB>
-where
-    Types: NodeTypesWithEngine,
-    DB: Database + DatabaseMetrics + DatabaseMetadata + Clone + Unpin + 'static,
-{
-    type DB = DB;
 }
 
 /// An adapter type that adds the builtin provider type to the user configured node types.
